+++ h3ron/src/iter.rs
@@ -9,6 +9,7 @@
 //!
 //! # Grid traversal
 //!
+//! * [`KRingBuilder`]
 //! * [`neighbors_within_distance_window_or_default`]
 //! * [`neighbors_within_distance_window`]
 //! * [`neighbors_within_distance`]
@@ -71,6 +72,104 @@ where
     }
 }
 
+/// `KRingBuilder` allows building k-rings with allocations only on the creation
+/// of the struct. This can be more efficient for large numbers of small (~ `k_max` <= 6) k-rings.
+/// (See the included `k_ring_variants` benchmark)
+///
+/// After calling [`KRingBuilder::build_k_ring`] the struct can be accessed
+/// as a [`Iterator`] returning `(H3Cell, u32)` tuples. The [`H3Cell`] value is a
+/// cell from within the k-ring and the `u32` distance `k` to the rings center.
+///
+/// TODO: find out why this method is slower for larger k-rings (see benchmark).
+pub struct KRingBuilder {
+    k_min: u32,
+    k_max: u32,
+
+    k_ring_indexes: Vec<H3Index>,
+    k_ring_distances: Vec<c_int>,
+    k_ring_size: usize,
+
+    current_pos: usize,
+}
+
+impl KRingBuilder {
+    /// `k_min` and `k_max` control the radius in which the neighbors will be iterated. Also
+    /// see [`H3Cell::k_ring`].
+    pub fn new(k_min: u32, k_max: u32) -> Self {
+        let k_ring_size = max_k_ring_size(k_max);
+
+        // pre-allocate the output vecs for k_ring_distances so we do not
+        // need to allocate during iteration.
+        let k_ring_indexes = vec![0; k_ring_size];
+        let k_ring_distances = vec![0; k_ring_size];
+        Self {
+            k_min,
+            k_max,
+            k_ring_indexes,
+            k_ring_distances,
+            k_ring_size,
+            current_pos: k_ring_size, // nothing left to iterate over
+        }
+    }
+
+    #[inline(always)]
+    fn rewind_iterator(&mut self) {
+        self.current_pos = 0;
+    }
+
+    /// Build the k-ring around the given [`H3Cell`]
+    ///
+    /// `k_min` and `k_max` control the radius in which the neighbors will be iterated. Also
+    /// see [`H3Cell::k_ring`].
+    ///
+    /// Building a k-ring resets the iterator to the start.
+    pub fn build_k_ring(&mut self, cell: &H3Cell) -> &mut Self {
+        // clear the pre-allocated vectors to ensure no values from the former run
+        // are left
+        unsafe {
+            // this essentially is `memset`
+            std::ptr::write_bytes(self.k_ring_indexes.as_mut_ptr(), 0, self.k_ring_size);
+            std::ptr::write_bytes(self.k_ring_distances.as_mut_ptr(), 0, self.k_ring_size);
+
+            // populate the pre-allocated vectors with the new neighbors
+            h3ron_h3_sys::kRingDistances(
+                cell.h3index(),
+                self.k_max as c_int,
+                self.k_ring_indexes.as_mut_ptr(),
+                self.k_ring_distances.as_mut_ptr(),
+            )
+        };
+        self.rewind_iterator();
+        self
+    }
+}
+
+impl Iterator for KRingBuilder {
+    type Item = (H3Cell, u32);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        while self.current_pos < self.k_ring_size {
+            let pos = self.current_pos;
+            self.current_pos += 1;
+
+            let h3index = self.k_ring_indexes[pos];
+            let k = self.k_ring_distances[pos] as u32;
+            if h3index == 0 || k < self.k_min {
+                // invalid h3index or `k` smaller the requested `k_min`,
+                // so it gets ignored.
+                continue;
+            }
+            return Some((H3Cell::new(h3index), k));
+        }
+        None
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        // IMPROVE: this overestimates when k_min != 0
+        (self.k_ring_size, None)
+    }
+}
+
 /// A `H3Cell` and one of its neighboring cells, combined with associated generic values.
 pub struct NeighborCell<'a, T> {
     /// The current cell of the iterator
@@ -104,14 +203,8 @@ pub struct CellNeighborsIterator<'a, I, F, T> {
     /// a neighbor.
     neighbor_default_value: Option<&'a T>,
 
-    /// The neighbors of the current cell which are still to visit.
-    next_neighbor_indexes: Vec<H3Index>,
-    next_neighbor_k: Vec<c_int>,
-    current_neighbor_i: usize,
-
-    k_min: u32,
-    k_max: u32,
-    k_ring_size: usize,
+    /// iterates over the neighbors of the current cell which are still to visit.
+    k_ring_builder: KRingBuilder,
 }
 
 /// See [`neighbors_within_distance_window_or_default`].
@@ -130,30 +223,17 @@ where
         loop {
             if let Some((cell, value)) = self.current_cell_key_value.as_ref() {
                 // advance until we find the next existing neighbor
-                while self.current_neighbor_i < self.k_ring_size {
-                    let this_current_neighbor_i = self.current_neighbor_i;
-                    self.current_neighbor_i += 1;
-                    if let (Some(neighbor_h3index), Some(neighbor_k)) = (
-                        self.next_neighbor_indexes.get(this_current_neighbor_i),
-                        self.next_neighbor_k.get(this_current_neighbor_i),
-                    ) {
-                        // invalid h3index or `k` smaller the requested `k_min`, so it gets ignored
-                        if (*neighbor_h3index == 0) || (*neighbor_k as u32) < self.k_min {
-                            continue;
-                        }
-
-                        let neighbor_cell = H3Cell::new(*neighbor_h3index);
-                        if let Some(neighbor_value) =
-                            (self.get_cell_value_fn)(&neighbor_cell).or(self.neighbor_default_value)
-                        {
-                            return Some(NeighborCell {
-                                cell: *cell,
-                                cell_value: *value,
-                                neighbor_cell,
-                                neighbor_value,
-                                k: *neighbor_k as u32,
-                            });
-                        }
+                while let Some((neighbor_cell, neighbor_k)) = self.k_ring_builder.next() {
+                    if let Some(neighbor_value) =
+                        (self.get_cell_value_fn)(&neighbor_cell).or(self.neighbor_default_value)
+                    {
+                        return Some(NeighborCell {
+                            cell: *cell,
+                            cell_value: *value,
+                            neighbor_cell,
+                            neighbor_value,
+                            k: neighbor_k,
+                        });
                     }
                 }
                 self.current_cell_key_value = None;
@@ -161,25 +241,7 @@ where
             if let Some(cell) = self.cell_iter.next() {
                 if let Some(cell_value) = (self.get_cell_value_fn)(cell.borrow()) {
                     self.current_cell_key_value = Some((*cell.borrow(), cell_value));
-
-                    // clear the pre-allocated vectors to ensure no values from the former run
-                    // are left
-                    self.next_neighbor_indexes
-                        .iter_mut()
-                        .for_each(|h3index| *h3index = 0);
-                    self.next_neighbor_k.iter_mut().for_each(|k| *k = 0);
-
-                    // populate the pre-allocated vectors with the new neighbors
-                    unsafe {
-                        h3ron_h3_sys::kRingDistances(
-                            cell.borrow().h3index(),
-                            self.k_max as c_int,
-                            self.next_neighbor_indexes.as_mut_ptr(),
-                            self.next_neighbor_k.as_mut_ptr(),
-                        )
-                    };
-                    // reset to start over from the beginning
-                    self.current_neighbor_i = 0;
+                    self.k_ring_builder.build_k_ring(cell.borrow());
                 }
             } else {
                 return None;
@@ -213,24 +275,12 @@ where
     F: Fn(&H3Cell) -> Option<&'a T>,
     F: 'a,
 {
-    let k_ring_size = max_k_ring_size(k_max);
-
-    // pre-allocate the output vecs for k_ring_distances so we do not
-    // need to allocate during iteration.
-    let next_neighbor_indexes = vec![0; k_ring_size];
-    let next_neighbor_k = vec![0; k_ring_size];
-
     CellNeighborsIterator {
         cell_iter,
         get_cell_value_fn,
         current_cell_key_value: None,
         neighbor_default_value,
-        next_neighbor_indexes,
-        next_neighbor_k,
-        current_neighbor_i: 0,
-        k_min,
-        k_max,
-        k_ring_size,
+        k_ring_builder: KRingBuilder::new(k_min, k_max),
     }
 }